backend-src/requirements.txt
fastapi==0.110.0
psycopg2_binary==2.9.9
uvicorn
